1.Request增加不用去重的field,去掉HelpUrl和TargetUrl注解,以接口方式发现需要加入爬取队列的url,加入MatchUrl注解,只用于判断当前抓取页面是否匹配该PageExtractor 2.Monitor 统计精度修改 3.增加站点爬取深度属性

+++ webmagic-extension/src/main/java/us/codecraft/webmagic/monitor/SpiderStatusMXBean.java
@@ -31,5 +31,5 @@ public interface SpiderStatusMXBean {
 
     public Date getStartTime();
 
-    public int getPagePerSecond();
+    public double getPagePerSecond();
 }
